refactor(server): 分析系サービスを services/analytics/ に集約
