+++ server/src/services/analytics/aggregate.ts
