+++ server/src/services/line-messaging.ts
@@ -9,7 +9,7 @@ import { stripMarkdown } from "~/lib/strip-markdown";
 import { createNeppChanAgent } from "~/mastra/agents/nepp-chan-agent";
 import { createRequestContext } from "~/mastra/request-context";
 import { injectBroadcastsToThread } from "~/services/broadcast-thread-injector";
-import { recordLlmUsage } from "~/services/llm-usage";
+import { recordLlmUsage } from "~/services/analytics/llm-usage";
 import { injectPollsToThread } from "~/services/poll-thread-injector";
 
 export const createLineClient = (token: string) =>
